backend/requirements.txt
Flask==2.1.3
google-cloud-vision==3.4.2
birdnetlib==0.18.0
librosa==0.9.2
werkzeug==2.1.2
